Add system-wide pip.conf including nexus as index
This is to make it easier to pip install packages from nexus in the
docker containers.

Note that the there is only a single copy of the `pip.conf` file in
the `docker/deploy` folder.  The rest are symbolic links, so we can
have consistent settings. The CI job conditions are updated to match.