Update PyTango from 9.3.1 to 9.3.2
With this new PyTango version, the symbolic link to the
libboost_python dynamic library does not have to be changed.

The tango-dsconfig installation of python-dsconfig is under Python 2.7,
and includes PyTango as a dependency.  However, PyTango isn't pinned
in its requirements, so we install it prior.  And we install numpy
before that, using the latest release that still supports Python 2.7.